+++ Netskope/netskope_modules/metrics.py
@@ -1,11 +1,18 @@
 from prometheus_client import Counter, Histogram
 
 # Declare prometheus metrics
-prom_namespace = "symphony_module_netskope"
+prom_namespace = "symphony_module_common"
 
 OUTCOMING_EVENTS = Counter(
     name="forwarded_events",
-    documentation="Number of events forwarded to SEKOIA.IO",
+    documentation="Number of events forwarded to Sekoia.io",
+    namespace=prom_namespace,
+    labelnames=["intake_key", "type"],
+)
+
+EVENTS_LAG = Histogram(
+    name="events_lags",
+    documentation="The delay, in seconds, from the date of the last event",
     namespace=prom_namespace,
     labelnames=["intake_key", "type"],
 )