+++ PagerDuty/tests/test_action_pagerduty_triggeralert.py
@@ -74,7 +74,7 @@ def test_pagerduty_postalert_default():
                 {
                     "src": "https://app.sekoia.io/assets/logos/sekoiaio-black.svg",
                     "href": "https://app.sekoia.io/",
-                    "alt": "SEKOIA.IO",
+                    "alt": "Sekoia.io",
                 }
             ],
             "links": [
@@ -84,6 +84,6 @@ def test_pagerduty_postalert_default():
                 }
             ],
             "event_action": "trigger",
-            "client": "SEKOIA.IO Security Service",
+            "client": "Sekoia.io Security Service",
             "client_url": "https://app.sekoia.io",
         }