+++ Salesforce/salesforce/connector.py
@@ -9,7 +9,7 @@ import orjson
 from aiolimiter import AsyncLimiter
 from dateutil.parser import isoparse
 from loguru import logger
-from pydantic import HttpUrl
+from pydantic import BaseModel, Field, HttpUrl
 from sekoia_automation.connector import Connector, DefaultConnectorConfiguration
 from sekoia_automation.module import Module
 from sekoia_automation.storage import PersistentJSON
@@ -20,19 +20,24 @@ from utils.file_utils import csv_file_as_rows, delete_file
 from .metrics import EVENTS_LAG, FORWARD_EVENTS_DURATION, OUTCOMING_EVENTS
 
 
-class SalesforceConfig(DefaultConnectorConfiguration):
-    """Configuration for SalesforceConnector."""
+class SalesforceModuleConfig(BaseModel):
+    """Configuration for SalesforceModule."""
 
+    client_secret: str = Field(secret=True)
     client_id: str
-    client_secret: str
     base_url: HttpUrl
-    ratelimit_per_minute: int = 60
 
 
 class SalesforceModule(Module):
-    """SalesforceConfig."""
+    """SalesforceModule."""
 
-    configuration: SalesforceConfig
+    configuration: SalesforceModuleConfig
+
+
+class SalesforceConnectorConfig(DefaultConnectorConfiguration):
+    """SalesforceConnector configuration."""
+
+    ratelimit_per_minute: int = 60
 
 
 class SalesforceConnector(Connector):
@@ -41,6 +46,7 @@ class SalesforceConnector(Connector):
     name = "SalesforceConnector"
 
     module: SalesforceModule
+    configuration: SalesforceConnectorConfig
 
     _salesforce_client: SalesforceHttpClient | None = None
 
@@ -88,7 +94,7 @@ class SalesforceConnector(Connector):
         if self._salesforce_client is not None:
             return self._salesforce_client
 
-        rate_limiter = AsyncLimiter(self.module.configuration.ratelimit_per_minute)
+        rate_limiter = AsyncLimiter(self.configuration.ratelimit_per_minute)
 
         self._salesforce_client = SalesforceHttpClient(
             client_id=self.module.configuration.client_id,
@@ -111,6 +117,8 @@ class SalesforceConnector(Connector):
         Returns:
             list[str]:
         """
+        logger.info("Pushing {count} events to intakes", count=len(events))
+
         return await asyncio.to_thread(
             self.push_events_to_intakes,
             events=events,
@@ -125,7 +133,7 @@ class SalesforceConnector(Connector):
             datetime: last event date
         """
         _last_event_date = self.last_event_date
-        log_files = await self.salesforce_client.get_log_files(_last_event_date.isoformat())
+        log_files = await self.salesforce_client.get_log_files(_last_event_date)
 
         logger.info(
             "Found {count} log files to process since {date}",
@@ -136,25 +144,41 @@ class SalesforceConnector(Connector):
         result = []
 
         for log_file in log_files.records:
-            log_file_date = isoparse(log_file.LogDate)
+            log_file_results = []
+            log_file_date = isoparse(log_file.CreatedDate)
             if _last_event_date < log_file_date:
                 _last_event_date = log_file_date
 
             records, csv_path = await self.salesforce_client.get_log_file_content(
-                log_file_uri=log_file.LogFile,
+                log_file=log_file,
             )
 
             if records is not None:
-                result.extend(await self._push_events([orjson.dumps(event).decode("utf-8") for event in records]))
+                log_file_results.extend(
+                    await self._push_events([orjson.dumps(event).decode("utf-8") for event in records])
+                )
 
             # Process csv file row by row to avoid memory issues
             if csv_path is not None:
                 async for row in csv_file_as_rows(csv_path):
-                    result.extend(await self._push_events([orjson.dumps(row).decode("utf-8")]))
+                    log_file_results.extend(await self._push_events([orjson.dumps(row).decode("utf-8")]))
 
                 await delete_file(csv_path)
 
+            logger.info(
+                "Finished to process log file {log_file_id}. Total amount of records is {count}",
+                log_file_id=log_file.Id,
+                count=len(log_file_results),
+            )
+
+            result.extend(log_file_results)
+
             with self.context as cache:
+                logger.info(
+                    "New last event date now is {last_event_date}",
+                    last_event_date=_last_event_date.isoformat(),
+                )
+
                 cache["last_event_date"] = _last_event_date.isoformat()
 
         return result
@@ -168,13 +192,13 @@ class SalesforceConnector(Connector):
             while True:
                 processing_start = time.time()
                 if previous_processing_end is not None:
-                    EVENTS_LAG.labels(intake_key=self.module.configuration.intake_key).observe(
+                    EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(
                         processing_start - previous_processing_end
                     )
 
                 message_ids: list[str] = loop.run_until_complete(self.get_salesforce_events())
                 processing_end = time.time()
-                OUTCOMING_EVENTS.labels(intake_key=self.module.configuration.intake_key).inc(len(message_ids))
+                OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(message_ids))
 
                 log_message = "No records to forward"
                 if len(message_ids) > 0:
@@ -183,7 +207,7 @@ class SalesforceConnector(Connector):
                 logger.info(log_message)
                 self.log(message=log_message, level="info")
 
-                FORWARD_EVENTS_DURATION.labels(intake_key=self.module.configuration.intake_key).observe(
+                FORWARD_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key).observe(
                     processing_end - processing_start
                 )