+++ Salesforce/tests/salesforce/test_connector.py
@@ -9,7 +9,7 @@ from aioresponses import aioresponses
 from sekoia_automation import constants
 
 from client.schemas.log_file import EventLogFile, SalesforceEventLogFilesResponse
-from salesforce.connector import SalesforceConnector, SalesforceModule
+from salesforce.connector import SalesforceConnector, SalesforceConnectorConfig, SalesforceModule
 
 
 @pytest.fixture
@@ -88,7 +88,13 @@ def pushed_events_ids(session_faker) -> list[str]:
 @pytest.fixture
 def connector(symphony_storage, client_id, client_secret, salesforce_url, pushed_events_ids, session_faker):
     module = SalesforceModule()
-    trigger = SalesforceConnector(module=module, data_path=symphony_storage)
+    config = SalesforceConnectorConfig(
+        intake_key=session_faker.word(),
+    )
+    trigger = SalesforceConnector(
+        module=module,
+        data_path=symphony_storage,
+    )
 
     # Mock the log function of trigger that requires network access to the api for reporting
     trigger.log = MagicMock()
@@ -98,12 +104,9 @@ def connector(symphony_storage, client_id, client_secret, salesforce_url, pushed
     trigger.push_events_to_intakes = MagicMock()
     trigger.push_events_to_intakes.return_value = pushed_events_ids
 
-    trigger.module.configuration = {
-        "client_id": client_id,
-        "client_secret": client_secret,
-        "base_url": salesforce_url,
-        "intake_key": session_faker.word(),
-    }
+    trigger.module.configuration = {"client_id": client_id, "client_secret": client_secret, "base_url": salesforce_url}
+
+    trigger.configuration = config
 
     return trigger
 
@@ -166,7 +169,7 @@ async def test_salesforce_connector_push_events_wrapper(connector, pushed_events
 
 @pytest.mark.asyncio
 async def test_salesforce_connector_get_salesforce_events(
-    connector, session_faker, http_token, csv_content, salesforce_url, pushed_events_ids
+    connector: SalesforceConnector, session_faker, http_token, csv_content, salesforce_url, pushed_events_ids
 ):
     """
     Test salesforce connector get salesforce events.
@@ -185,7 +188,8 @@ async def test_salesforce_connector_get_salesforce_events(
         EventType=session_faker.pystr(),
         LogFile=session_faker.pystr(),
         LogDate=log_file_date.isoformat(),
-        LogFileLength=session_faker.pyfloat(),
+        CreatedDate=log_file_date.isoformat(),
+        LogFileLength=len(csv_content.encode("utf-8")),
     )
 
     log_files_response_success = SalesforceEventLogFilesResponse(totalSize=1, done=True, records=[event_log_file])
@@ -200,7 +204,7 @@ async def test_salesforce_connector_get_salesforce_events(
             payload=token_data,
         )
 
-        query = connector.salesforce_client._log_files_query(connector.last_event_date.isoformat())
+        query = connector.salesforce_client._log_files_query(connector.last_event_date)
         get_log_files_url = connector.salesforce_client._request_url_with_query(query)
 
         log_file_response_dict = log_files_response_success.dict()
@@ -225,7 +229,7 @@ async def test_salesforce_connector_get_salesforce_events(
 
 @pytest.mark.asyncio
 async def test_salesforce_connector_get_salesforce_events_1(
-    connector, session_faker, http_token, csv_content, salesforce_url, pushed_events_ids
+    connector: SalesforceConnector, session_faker, http_token, csv_content, salesforce_url, pushed_events_ids
 ):
     """
     Test salesforce connector get salesforce events.
@@ -249,7 +253,8 @@ async def test_salesforce_connector_get_salesforce_events_1(
         EventType=session_faker.pystr(),
         LogFile=session_faker.pystr(),
         LogDate=log_file_date.isoformat(),
-        LogFileLength=session_faker.pyfloat(),
+        CreatedDate=log_file_date.isoformat(),
+        LogFileLength=1024 * 1024 * 1024,
     )
 
     log_files_response_success = SalesforceEventLogFilesResponse(totalSize=1, done=True, records=[event_log_file])
@@ -264,7 +269,7 @@ async def test_salesforce_connector_get_salesforce_events_1(
             payload=token_data,
         )
 
-        query = connector.salesforce_client._log_files_query(connector.last_event_date.isoformat())
+        query = connector.salesforce_client._log_files_query(connector.last_event_date)
         get_log_files_url = connector.salesforce_client._request_url_with_query(query)
 
         log_file_response_dict = log_files_response_success.dict()