+++ Sophos/sophos_module/trigger_sophos_xdr_query.py
@@ -6,6 +6,7 @@ from functools import cached_property
 
 import orjson
 from requests.exceptions import HTTPError
+from urllib3.exceptions import HTTPError as BaseHTTPError
 from sekoia_automation.connector import DefaultConnectorConfiguration
 from sekoia_automation.storage import PersistentJSON
 
@@ -17,6 +18,7 @@ from sophos_module.metrics import INCOMING_EVENTS, OUTCOMING_EVENTS, FORWARD_EVE
 
 class SophosXDRQueryConfiguration(DefaultConnectorConfiguration):
     chunk_size: int = 1000
+    frequency: int = 60
 
 
 class SophosXDRQueryTrigger(SophosConnector):
@@ -75,27 +77,44 @@ class SophosXDRQueryTrigger(SophosConnector):
         self.log(message="Sophos Events Trigger has started", level="info")
 
         try:
-            start = time.time()
+            while self.running:
+                start = time.time()
+                try:
+                    self.getting_results(self.pagination_limit)
 
-            self.getting_results(self.pagination_limit)
+                except HTTPError | BaseHTTPError as ex:
+                    self.log_exception(ex, message="Failed to get next batch of events")
+                except Exception as ex:
+                    self.log_exception(ex, message="An unknown exception occurred")
+                    raise
 
-            duration = int(time.time() - start)
-            FORWARD_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key).observe(duration)
+                # compute the duration of the last events fetching
+                duration = int(time.time() - start)
+                FORWARD_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key).observe(duration)
 
-        except HTTPError as ex:
-            self.log_exception(ex, message="Failed to get next batch of events")
-        except Exception as ex:
-            self.log_exception(ex, message="An unknown exception occurred")
-            raise
+                # Compute the remaining sleeping time
+                delta_sleep = self.configuration.frequency - duration
+                # if greater than 0, sleep
+                if delta_sleep > 0:
+                    time.sleep(delta_sleep)
+        finally:
+            self.log(message="Sophos Events Trigger has stopped", level="info")
 
     def post_query(self, query: dict):
         # It's the first step of our query treatment.
         # Posting the query
+        self.log(
+            message=f"Querying the Sophos data lake at {self.from_date.strftime('%Y-%m-%d %H:%M:%S')}", level="info"
+        )
         response_runQuery = self.client.run_query(json_query=query).json()
 
         query_id = response_runQuery.get("id")
 
+        if query_id == None:
+            return "failed", query_id
+
         status = ""
+        self.log(message=f"Waiting the query {query_id} to complete", level="info")
         while status != "finished":
             # Give time for query to finish.
             time.sleep(2)
@@ -105,46 +124,52 @@ class SophosXDRQueryTrigger(SophosConnector):
             status = response_queryStatus.get("status")
             result = response_queryStatus.get("result")
 
+        self.log(message=f"Finishing the loop with status {status} and result {result}", level="info")
+
         return result, query_id
 
     def getting_results(self, pagination: str):
         now = datetime.now(timezone.utc)
+
         result, query_id = self.post_query(self.query)
 
         if result != "succeeded":
-            raise Exception("Sophos request is correct and sent treatment but failed!!")
-
-        # If the result succeed ==> get the data.
-        response = self.client.get_query_results(query_id, pagination).json()
-        items = response.get("items", [])
-        INCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(items))
-
-        if len(items) > 0:
-            messages = [orjson.dumps(message).decode("utf-8") for message in items]
-            nextKey = response.get("pages").get("nextKey")
-            self.log(message=f"Sending the first batch of {len(messages)} messages", level="info")
-            OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(messages))
-            self.push_events_to_intakes(events=messages)
-            self.events_sum += len(messages)
-
-            while nextKey:
-                grouped_data = []
-                response_next_page = self.client.get_query_results_next_page(query_id, pagination, nextKey).json()
-                next_page_items = response_next_page.get("items", [])
-                grouped_data.extend(next_page_items)
-                self.log(message=f"Sending other batches of {len(grouped_data)} messages", level="info")
-                OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(grouped_data))
-                self.push_events_to_intakes(events=grouped_data)
-                self.events_sum += len(grouped_data)
-                nextKey = response_next_page.get("pages").get("nextKey")
-
-            most_recent_date_seen = now
-            self.from_date = most_recent_date_seen
-            with self.context as cache:
-                cache["most_recent_date_seen"] = most_recent_date_seen.isoformat()
-
+            self.log(
+                message=f"The query failed. No events collected at {self.from_date.strftime('%Y-%m-%d %H:%M:%S')} !!"
+            )
         else:
-            self.log(message="No messages to forward", level="info")
+            # If the result succeed ==> get the data.
+            response = self.client.get_query_results(query_id, pagination).json()
+            items = response.get("items", [])
+            self.log(message=f"Getting results with {len(items)} elements", level="info")
+            INCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(items))
+
+            if len(items) > 0:
+                messages = [orjson.dumps(message).decode("utf-8") for message in items]
+                nextKey = response.get("pages").get("nextKey")
+                self.log(message=f"Sending the first batch of {len(messages)} messages", level="info")
+                OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(messages))
+                self.push_events_to_intakes(events=messages)
+                self.events_sum += len(messages)
+
+                while nextKey:
+                    grouped_data = []
+                    response_next_page = self.client.get_query_results_next_page(query_id, pagination, nextKey).json()
+                    next_page_items = response_next_page.get("items", [])
+                    grouped_data.extend(next_page_items)
+                    self.log(message=f"Sending other batches of {len(grouped_data)} messages", level="info")
+                    OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(grouped_data))
+                    self.push_events_to_intakes(events=grouped_data)
+                    self.events_sum += len(grouped_data)
+                    nextKey = response_next_page.get("pages").get("nextKey")
+
+                most_recent_date_seen = now
+                self.from_date = most_recent_date_seen
+                with self.context as cache:
+                    cache["most_recent_date_seen"] = most_recent_date_seen.strftime("%Y-%m-%dT%H:%M:%SZ")
+
+            else:
+                self.log(message="No messages to forward", level="info")
 
 
 class SophosXDRIOCQuery(SophosXDRQueryTrigger):
@@ -152,5 +177,5 @@ class SophosXDRIOCQuery(SophosXDRQueryTrigger):
         super().__init__(*args, **kwargs)
         self.query = {
             "adHocQuery": {"template": "SELECT * FROM xdr_ioc_view WHERE ioc_detection_weight > 3"},
-            "from": self.from_date.isoformat(),
+            "from": self.from_date.strftime("%Y-%m-%dT%H:%M:%SZ"),
         }