+++ VadeCloud/vadecloud_modules/__init__.py
@@ -1,0 +1,7 @@
+from sekoia_automation.module import Module
+
+from .models import VadeCloudModuleConfiguration
+
+
+class VadeCloudModule(Module):
+    configuration: VadeCloudModuleConfiguration