+++ Sophos/CHANGELOG.md
@@ -7,6 +7,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.11.9] - 2023-07-26
+
+### Changed
+
+- fix metrics
+
+## [1.11.8] - 2023-07-24
+
+### Changed
+
+- Fix the way to handle exception in the EDR connector
+- Add more logs
+
+## [1.11.7] - 2023-07-18
+
+### Added
+
+- add more logs
+- Add an interval between two requests in the XDR connector
+- Improve the way to handle failing requests
+
+## [1.11] - 2023-07-13
+
+### Changed
+
+- Bump sdk version
+
+## [1.10] - 2023-07-12
+
+### Changed
+
+- Bump sdk version
+
 ## [1.9.1] - 2023-06-30
 
 ### Added